pyproject.toml
[project]
name = "realtime-sentiment-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alpaca-trade-api==3.2.0",
    "fastapi==0.115.12",
    "ipykernel==6.29.5",
    "python-dotenv==1.1.0",
    "torch==2.6.0",
    "transformers==4.51.2",
    "uvicorn==0.34.0",
    "websockets==10.4",
]
